fixed sync issues
